requirements-langchain.txt
ipywidgets
python-dotenv==1.0.0
langchain==0.0.137
pinecone-client==2.2.1
openai
tiktoken
streamlit
wikipedia
chromadb
pypdf2
faiss-cpu
sentence_transformers
watchdog
accelerate